+++ packages/mockserver-config-writer/test/unit/mockserver-config/ui5-mock-yaml.test.ts
@@ -12,9 +12,51 @@ describe('Test enhanceYaml()', () => {
     const webappPath = join('/webapp');
     const manifestJsonPath = join(webappPath, 'manifest.json');
     const manifestWithMainService = `{"sap.ui5": { "models": { "": { "dataSource": "ds" } } },"sap.app": { "dataSources": { "ds": { "uri": "ds/uri/" } } }}`;
-
+    const mockManifestJson = `{
+        "sap.app": {
+            "id": "mockserverv2",
+            "dataSources": {
+                "mainService": {
+                    "uri": "/sap/opu/odata/sap/SEPMRA_PROD_MAN/"
+                },
+                "SEPMRA_PROD_MAN": {
+                    "uri": "/sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='SEPMRA_PROD_MAN',Version='0001')/$value/",
+                    "type": "ODataAnnotation",
+                    "settings": {
+                        "localUri": "localService/SEPMRA_PROD_MAN.xml"
+                    }
+                },
+                "annotation": {
+                    "type": "ODataAnnotation",
+                    "uri": "annotations/annotation.xml",
+                    "settings": {
+                        "localUri": "annotations/annotation.xml"
+                    }
+                }
+            }
+        }
+    }`;
     beforeEach(() => {
         jest.clearAllMocks();
+    });
+
+    test('Create new ui5-mock.yaml with annotations from mock manifest.json', async () => {
+        const fs = getFs({ [manifestJsonPath]: mockManifestJson });
+        await enhanceYaml(fs, basePath, webappPath, { path: '/path/for/new/config' });
+
+        expect(fs.read(ui5MockYamlPath)).toMatchSnapshot();
+
+        const ui5Config = await UI5Config.newInstance(fs.read(ui5MockYamlPath));
+        const mockserverConfig = ui5Config.findCustomMiddleware<MockserverConfig>('sap-fe-mockserver');
+        expect(mockserverConfig?.configuration.services?.[0].urlPath).toBe('/path/for/new/config');
+        expect(mockserverConfig?.configuration.annotations).toEqual([
+            {
+                localPath: './webapp/localService/SEPMRA_PROD_MAN.xml',
+                urlPath:
+                    "/sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='SEPMRA_PROD_MAN',Version='0001')/$value/"
+            },
+            { localPath: './webapp/annotations/annotation.xml', urlPath: 'annotations/annotation.xml' }
+        ]);
     });
 
     test('Update ui5-mock.yaml, path from manifest', async () => {
@@ -23,10 +65,37 @@ describe('Test enhanceYaml()', () => {
         expect(fs.read(ui5MockYamlPath)).toMatchSnapshot();
     });
 
+    test('Update ui5-mock.yaml, path from manifest with annotations', async () => {
+        const fs = getFsWithUi5MockYaml(mockManifestJson);
+        await enhanceYaml(fs, basePath, webappPath);
+        expect(fs.read(ui5MockYamlPath)).toMatchSnapshot();
+    });
+
     test('Update old ui5-mock.yaml with given path', async () => {
         const fs = getFsWithUi5MockYaml('{}');
         await enhanceYaml(fs, basePath, webappPath, { path: 'path/to/service' });
         expect(fs.read(ui5MockYamlPath)).toMatchSnapshot();
+    });
+
+    test('Create new ui5-mock.yaml based on ui5.yaml, updated with annotations', async () => {
+        const fs = getFsWithUi5Yaml(mockManifestJson);
+        await enhanceYaml(fs, basePath, webappPath, { path: 'new/path/to/service' });
+        expect(fs.read(ui5MockYamlPath)).toMatchSnapshot();
+        // additional check of urlPath, even if snapshot test get lightheartedly updated, the urlPath should remain stable.
+        const ui5Config = await UI5Config.newInstance(fs.read(ui5MockYamlPath));
+        expect(
+            ui5Config.findCustomMiddleware<MockserverConfig>('sap-fe-mockserver')?.configuration.services?.[0].urlPath
+        ).toBe('new/path/to/service');
+        expect(
+            ui5Config.findCustomMiddleware<MockserverConfig>('sap-fe-mockserver')?.configuration.annotations
+        ).toEqual([
+            {
+                localPath: './webapp/localService/SEPMRA_PROD_MAN.xml',
+                urlPath:
+                    "/sap/opu/odata/IWFND/CATALOGSERVICE;v=2/Annotations(TechnicalName='SEPMRA_PROD_MAN',Version='0001')/$value/"
+            },
+            { localPath: './webapp/annotations/annotation.xml', urlPath: 'annotations/annotation.xml' }
+        ]);
     });
 
     test('Create new ui5-mock.yaml based on ui5.yaml', async () => {