+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sap-ux/cap-config-writer
 
+## 0.2.24
+
+### Patch Changes
+
+-   Updated dependencies [87c942e5]
+    -   @sap-ux/project-access@1.19.4
+
+## 0.2.23
+
+### Patch Changes
+
+-   Updated dependencies [89013210]
+    -   @sap-ux/project-access@1.19.3
+
+## 0.2.22
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.2
+
+## 0.2.21
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.1
+
 ## 0.2.20
 
 ### Patch Changes