+++ packages/ui5-library-writer/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @sap-ux/ui5-library-writer
 
+## 0.3.24
+
+### Patch Changes
+
+-   Updated dependencies [87c942e5]
+    -   @sap-ux/project-access@1.19.4
+    -   @sap-ux/ui5-application-writer@0.26.5
+
+## 0.3.23
+
+### Patch Changes
+
+-   Updated dependencies [89013210]
+    -   @sap-ux/project-access@1.19.3
+    -   @sap-ux/ui5-application-writer@0.26.5
+
+## 0.3.22
+
+### Patch Changes
+
+-   Updated dependencies [ec509c40]
+    -   @sap-ux/ui5-config@0.22.0
+    -   @sap-ux/project-access@1.19.2
+    -   @sap-ux/ui5-application-writer@0.26.5
+
+## 0.3.21
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.1
+-   @sap-ux/ui5-application-writer@0.26.4
+
+## 0.3.20
+
+### Patch Changes
+
+-   4b29ddcc: Update TypeScript templates, eslint config and ui5 devDependencies
+-   Updated dependencies [4b29ddcc]
+    -   @sap-ux/ui5-application-writer@0.26.4
+
 ## 0.3.19
 
 ### Patch Changes